requirements.txt
aiogram>=3.0
yookassa
python-dotenv
sqlalchemy>=2.0
aiosqlite
pytz
